requirements.txt
alembic==1.6.4
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
dlib==19.22.99
dnspython==2.1.0
email-validator==1.1.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.1
imageio==2.9.0
ipython==7.23.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.52
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
SQLAlchemy==1.4.15
threadpoolctl==2.1.0
tifffile==2021.4.8
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
